requirements.txt
opencv-python
face-recognition
face-recognition-models
numpy
pystray
Pillow
plyer
win10toast
